requirements.txt
asgiref==3.5.2
backports.zoneinfo==0.2.1
Django==4.0.6
djangorestframework==3.13.1
djangorestframework-api-key==2.2.0
djangorestframework-simplejwt==5.2.0
importlib-metadata==4.12.0
Markdown==3.3.7
postgres==4.0
psycopg==3.0.15
psycopg2-binary==2.9.3
psycopg2-pool==1.1
PyJWT==2.4.0
pytz==2022.1
sqlparse==0.4.2
zipp==3.8.1
python-dotenv~=0.20.0